feat/SEOULDATA-66 rename: ReviewController 위치 수정
